+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 ## is.codion.framework.model
 - EntitySearchModel.Builder.attributes() added.
 - EntitySearchModel.Builder.orderBy() added.
+- EntitySearchModel.entityType() replaced with entityDefinition().
 ### is.codion.swing.common.ui
 - FilterTableCellRenderer.UISettings.selectionForeground() added.
 - DefaultFilterTableCellRenderer uses UISettings.selectionForeground().
@@ -15,9 +16,12 @@ Codion Change Log
 - LookAndFeelSelectionDialogBuilder.includePlatformLookAndFeels() added along with INCLUDE_PLATFORM_LOOK_AND_FEELS, for excluding platform look and feels from the selection if auxiliary ones are available.
 - Components.textField() no longer forwards call to NumberField in case of numerical values.
 - ColumnConditionPanel.FieldFactory.createLowerField() and createUpperField() now with default implementations.
+### is.codion.swing.framework.model
+- EntityComboBoxModel.entityType() replaced with entityDefinition().
 ### is.codion.swing.framework.ui
 - EntityComponents.textField() now forwards call to NumberField in case of numerical attributes.
 - EntityComponents.textField() now creates an non-focusable, read only text field for item based attributes.
+- EntitySearchField.selectorFactory now final, selectorFactory() removed.
 
 ## 0.18.24
 ## is.codion.common.core